Fixed: Add elvis/elvis/media/ to .gitignore (with .gitkeep)
In the previous commit, I removed all the files in this directory. This will be
the MEDIA_ROOT when PRODUCTION is False, so I want the directory to exist (thus
the addition of .gitkeep), but I don't want to commit anything that appears in
there (thus the inclusion in .gitignore).