Updated: Base, temporary uploads template
Changed some of the JS on the temporary uploads template to more accurately reflect what it should do.

Pills in base template removed, and navigation is now done entirely through the nav bar. Not a final decision, but will give me more space to work with facets in the search template.