Added: Facets for searching
Facets now work and are logically joined by default using AND. This is because the way a user interacts with facets is by sequentially clicking on the sidebar. Each facet clicked should be required in the results.

On the other hand, if they want to limit results using OR, they can do so in the advanced search.