Fixed: Facet Ordering
After searching through the documentation and trying many different things, I think the only way of getting the facet fields sorted consistently is to have the facets in a list of tuples and not a dict. The sorting is quite hacky and definitely slows down the query process, but I don't see any other way of doing it.

Fixes #11