New: Created views.py file for non-model views
Currently we have a view/forms.py file which holds a number of half-compelted views for which perform general/utility functions and do not directly correspond to a model. Since these forms are only half working (and not referenced anywhere I can see), I've made a file with a similar function, with a better name (I think). I've moved the solr_suggest method into this file.