New: Added views to upload and unzip files.
The **upload_files** view iterates through all the files in a request's request.FILES and uploads those with acceptable extensions to the media/temp directory. It also ignores bad-prefixes (in settings.py) like '.' and '__' that could mess with the os. It returns a dict containing the name, creator, and path of all the files (maybe the path is useless/obvious, but I was thinking of directories inside zip files when I put it in - which I don't think we should support). If an uploaded file is a zip, it will pass it to **unzip_file** and add the extracted files to the returned dict before deleting the archive.

The **unzip_file** view will *only* extract files with an acceptable extension. It returns a list of the file names which were extracted to the directory it takes as an arg.

**settings-example.py** has had the ELVIS_BAD_PREFIX field added.