New: Added multiple input awareness to autocomplete.
Now if the field can have multiple inputs, the autocomplete function can play nice with it. That is, it can add suggestions from the menu to the list, and will only offer suggestions for the last item in the list. I'm actually kinda surprised I was able to add this in half an hour and it seems to work fine. :)