New: Added organized tabs to the piece create template.
Broke up the input fields into groups and placed them on different tabs. Added a button at the bottom of the page which either moves to the next tab or submits the form, based on context. Added instructions on how to write lists for the fields which can have multiple inputs.