WIP: Added cleanup class.
An instance of this class can be passed around all the creation methods and it holds all created objects in a list. If something goes sideways at any point, the classes cleanup method is called before re-raising the exception. This way, if a piece-create fails for any reason, the database won't be filled with garbage.