Fixed: Composer and Collection lists no longer slow!
The composer and collection models where using the same serializers for detail and list view. This meant that viewing a list of 10 composers was serving all the pieces and movements associated with the 10 composers... Potentially thousands of objects where being served up for a list with 10 things. Now it doesn't do that, and instead just serves up the integer of how many pieces are there.