New: migrate_tags in working state.
The function iterates through the csv and creates/attaches all models (which are currently tags) to pieces and models. It also has a delete function that can be run afterwards to delete tags which are no longer attached to anything (which should be all the non-free tags in the csv). Also added the 'religiosity' field to movements and pieces, which, since there are only going to be four states (Unknown, Ambiguous, Secular, Sacred), I decided was easier to have as a text field rather than a model in itself.

The migrate_tags function should be called from inside a django shell.