WIP: Composer detail page
This page will now have a list of pieces and a list of free movements: that is, movements that are not part of a piece. This is an improvement, as before it loaded all pieces and all movements attached to the composer - producing a list of hundreds of movements (often with the same names) that had no context. The pages load faster and make more sense now.