Fixed: Models can now be imported from django db-dump.
Checking the 'raw' kwarg (which is true during a `manage.py loaddata` call) and ignoring the post-save (if raw) allows the database to be dumped to a json file and loaded by Django. This was usefull in getting the database on to my laptop AND may be usefull in the future.