New: Rewrote and improved search-backend.
Now all text fields can recieve as input boolean statements that contain AND, OR, NOT and brackets, which will be interpreted correctly by solr. The fields in advanced search act as filters for the query, that is, their values are OR'd together. Facets are AND'd together, allowing to 'drill' for more specific results that meet multiple criteria.