WIP: Downloads fixed.
Now javascript will only draw a link on the page when it has confirmed the zipped file is done. Also, this link is a direct link to the file as it should stand in /media/.

When a zip file is created for a user download, it will now also call its own deletion with a 10 minute delay. This is much more practical than running an entire other celery script which just deletes all user download files at midnight.