New: Piece creation/updating form error reporting.
Now the form error printout will print the aria-label of the input rather than the id, which is much more human-readable.
A lack of sources on file inputs will now return form errors - but only if there is actually a file in the upload-field.
Closing the 'form errors' modal now scrolls to the first instance of an error in the form.