Fixed: Shared 'submit' button now behaved in same way.
Now that the piece form is being shared between upload and update, the submit button needs to behave in the same way. The update form was relying on it being a 'type=button', while the upload form was relying on it being a 'type=submit'. This should normalize the behaviour.