New: Added attached_files solr field.
Now pieces and movements will also index the paths of the files that are attached to them.
Also included in this commit are some deduping stuff on the django side (it was simpler than setting up a solr deduper in this case), and some new methods for Piece and Movement that allow you to index the item in solr without writing to the DB, and optionally without commiting to solr (useful for batch re-indexing, which I'm about to do).