Fixed: Multiple searching issues.
Multiple file formats as filters will now be OR'd together - so saying you're interested in .midi and .mei widens rather than narrows the search. Fixes #126.
Selecting multiple tag facets are now properly formatted and won't break the search. Tags are the only field where selecting multiple facets is possible, so it was a bit of a special case. Fixes #155