WIP: Re-wrote most common methods for session based cart.
This cart will be WAY faster than the current cart. The current cart does all its atomic actions (adding/removing a single item) as database updates. The goal on this branch is to move the entire site to a session based cart system. Not only will it avoid lots of database reads/writes, it also benefits from storing the entire cart as a dict, which will eliminate a large number of linear searches which were occuring on various actions.