Fixed: Removed timeout on cart caches.
They should stay in cache until they are written through by a
save/delete of the model in question. Note: This implies that
the memcaching software (redis) should have an upper limit to
the amount of information it will cache (otherwise you might fill
your entire ram with cached stuff). I have this set at 200MB
locally, and I imagine that is probably enormous for what is
being cached.