WIP: Updated Attachment Serializers.
I'm planning to go through the whole project and reformat all the serializers so there is a 'full' and 'min' (and perhaps a 'list' for those that need it) serializer for each model, which they can use on eachother. This will really simplify things, rather than having each serializer type writing its own special serializers for each model it needs to reference.