New: Added unique uuid's to all models.
This turned out to be a tricky change - adding unique fields to existing models requires 3 migrations, and the solr database had to be configured for this change as well, and everything needed to be re-indexed.

The upshots of this is that every model now has a unique site-wide uuid. This will make it much easier to refer to specific models without ambiguity, cache serialization results, and pass around references to models that can be easily dereferenced at any time.