WIP: Began writing common view behaviour.
It seems like the easiest way to allow only the creator/superusers to edit/delete individual objects is to simply check the creator on each request to a modification. Object-level permission level plugins exist, but since each object already has a creator, I have everything needed to implement object-level permissions myself.

The goal of creating common view superclasses is to standardize the behaviour of all views and facilitate the transition to a more restful API.