WIP: Added common listcreate view.
The listcreate view will allow query parameters to be passed in in order to sort/filter the list being shown. I already do something like this in order to display the collections/piecs that belong to a specific user (under the 'my pieces' button), but this superclass will add this standardized behaviour to all lists.