New: Standardized all list-style templates.
I think building up the templates from smalled object-like templates is a really great way to make the site easy to update and uniform. Here I've standardized the body and entries in list-style templates. All the individual templates need to do is pass which of their context variables they want to show in the list into the base_list_item template. I would like to do this with the detail templates as well.