New: Added media view.
The Django manual said the way it had been written was 'grossly inefficient and probably dangerous'. So now its an authentication protected view with a redirect to login.

I should write something here at some point to serve a more descriptive message as to why the redirect is happening, for people looking at the json.