Fixed: Download buttons on search page.
Added a `force_redraw` optional flag to init_cart_button() for the search page to use. This flag forces all cart_buttons to be rediscovered and drawn again. This is nice because now the search page and the rest of the site use the same script for cart button behaviour.