New: Automatic file renaming.
Attachments have a new function called auto_rename that will check if the Attachment should rename itself, and apply the renaming if so. This renaming will carry over to the file on the OS.

Pieces and Movements will now call their Attachments auto_rename() function after they are saved.
Composer was given a rename() function that will rename all its attachments. It would be crazy inefficient to check if tens/hundreds/thousands of files need to be renamed every time a compose is saved.

#Fixes #171