New: File name normalization.
Moved the normalizer into its own helper file (since it's used quite often in the project).

The Attachment.auto_rename() function was changed to use this new normalizer. It was also improved so that files under movements will attempt to find the piece above the movement, and put both names into the file.