New: All user specific data determined at serialization.
Moved the permission checks in the common detail views GET method to the serializers. The previous setup would need to find each object in a GET request twice, once to serialize it, and again to check a user's permissions. By doing both at serialization time, we are being much more efficient.

Also renamed and doc-string'd the methods in the serializer mixin I've created to do handle user specific data in serialization responses.