WIP: Solr 5 switch.
Switched configs over to solr5. We'll have multiple cores for the dev/prod branches, and a core for testing on. The test core has its config files symlinked from the dev core.

To set this up with solr5, make a symlink in /opt/solr/server/solr to $REPO_ROOT/solr/elvis_dev.