New: Can request specific values from endpoints.
Now, all the endpoints that serve a specific type of object can be passed in query parameters to limit which values actually get returned. This is intended to make ajax calls that need specific information from a view faster.

The list of values you are interested in should be passed in ?values[]=x&values[]=y...