Fixed: Tests use a test solr core.
This is helpful for isolating indexing issues and tests on solr. The solr test core is a copy of the solr dev core, so should act exactly the same, but gives the flexibility to start from a clean slate and perform extensive tests on the solr conf without destroying our dev environments.