New: Added generate_file_report command.
This new management command produces a report on the files
in the database. It generates a list of files that are missing,
and organizes files into groups if they are duplicates of
each other. The report is printed to the screen in json,
use the shell to save it as a file.