WIP: TOS feature.
The TOS feature requires a number of new components to implement:

- An additional UserProfile class to add a flag to the User stating
whether they have accepted the TOS or not. A migration has been
set up to create a UserProfile for each existing User.
- New middleware to redirect authenticated GET requests to the
TOS screen if the user has no yet accepted TOS.
- A new TOS view and template.

This is a WIP commit. There are a number of steps left to do:
1. Finish the TOS page template with a form to accept.
2. Write the POST method for the TOS page view class.
3. Set up a listener or method to create a UserProfile each time
a new user is created.