Fixed: UserProfile created with Users.
Updated the UserForm to save a UserProfile along with any new
User it creates. Hopefully this is a better solution than simply
adding this behaviour to the register view or in a signal handler.

As long as the UserForm is used to create Users, we should always
have it be the case that it will create a User and UserProfile when
a call to save() succeeds.