Fixed: Creating related models before Attachments.
Lots of code duplication in the Attachments model is related
to the fact that if supporting models (such as parent piece and
composer) do not exist at the time a file is attached to an
Attachment, you have no notion of what to name the file.

This duplication can be removed by ensuring that supporting models
are all created before Attachments. This is how it should be in the
first place, since an Attachment depends on multiple things yet
nothing depends on an Attachment.